requirements.txt
redis
ulid-py
pytest
